clean-files and more instructions for certifier_test

+++ vm_model_tools/examples/scenario1/clean-files.sh
@@ -11,18 +11,37 @@
 # --------------------------------------------------------------------------------
 
 echo "Removing runtime data files"
-rm policy_store.dom0* cryptstore.dom0* quote_cert.crt ekchain.bin
-pushd provisioning
-  rm ./*
-popd
- pushd service
- rm ./*
-popd
-pushd cf_data
- rm ./*
-popd
+if [[ ! -e "$EXAMPLE_DIR" ]] ; then
+  echo "EXAMPLE_DIR missing or not set"
+  exit
+fi
 
-rm measurement cf_utility.measurement sev_cf_utility.measurement
+pushd $EXAMPLE_DIR
+  rm policy_store.dom0* cryptstore.dom0* quote_cert.crt ekchain.bin
+  rm measurement cf_utility.measurement sev_cf_utility.measurement
+
+  if [[ ! -e "./provisioning" ]] ; then
+    echo "provisioning directory missing or not set"
+    exit
+  fi
+  pushd provisioning
+    rm ./*
+  popd
+  if [[ ! -e "./service" ]] ; then
+    echo "service directory missing or not set"
+    exit
+  fi
+  pushd service
+   rm ./*
+  popd
+  if [[ ! -e "./cf_data" ]] ; then
+    echo "cf_data directory missing or not set"
+    exit
+  fi
+  pushd cf_data
+    rm ./*
+  popd
+popd
 
 echo "Removed runtime data files"
